common:car has it's own namespace now Renamed common:car package from auto to car Removed google-services plugin
