+++ androidApp/src/main/AndroidManifest.xml
@@ -55,7 +55,7 @@
         </provider>
 
         <service
-                android:name=".auto.ConfettiCarAppService"
+                android:name=".car.ConfettiCarAppService"
                 android:exported="true">
             <intent-filter>
                 <action android:name="androidx.car.app.CarAppService"/>
@@ -74,7 +74,7 @@
             </intent-filter>
         </activity>
 
-        <activity android:name=".auto.signin.SignInWithGoogleActivity" />
+        <activity android:name=".car.signin.SignInWithGoogleActivity" />
 
     </application>
 </manifest>
