+++ common/car/build.gradle.kts
@@ -5,7 +5,6 @@ import java.util.*
 plugins {
     id("com.android.library")
     kotlin("android")
-    id("com.google.gms.google-services")
 }
 
 configureCompilerOptions()
@@ -36,7 +35,7 @@ android {
         kotlinCompilerExtensionVersion = libs.versions.compose.compiler.get()
     }
 
-    namespace = "dev.johnoreilly.confetti"
+    namespace = "dev.johnoreilly.confetti.car"
 }
 
 dependencies {
