--- common/car/src/main/java/dev/johnoreilly/confetti/auto/utils/DecomposeUtils.kt
@@ -1,11 +1,0 @@
-package dev.johnoreilly.confetti.auto.utils
-
-import androidx.car.app.Screen
-import com.arkivanov.decompose.ComponentContext
-import com.arkivanov.decompose.DefaultComponentContext
-import com.arkivanov.essenty.lifecycle.essentyLifecycle
-
-fun Screen.defaultComponentContext(): ComponentContext =
-    DefaultComponentContext(
-        lifecycle = essentyLifecycle(),
-    )
