+++ common/car/src/main/java/dev/johnoreilly/confetti/car/ui/ErrorScreen.kt
@@ -1,0 +1,32 @@
+package dev.johnoreilly.confetti.car.ui
+
+import androidx.annotation.StringRes
+import androidx.car.app.CarContext
+import androidx.car.app.Screen
+import androidx.car.app.model.Action
+import androidx.car.app.model.CarIcon
+import androidx.car.app.model.MessageTemplate
+import androidx.car.app.model.Template
+import dev.johnoreilly.confetti.car.R
+
+class ErrorScreen(
+    carContext: CarContext,
+    @StringRes val resId: Int,
+) : Screen(carContext) {
+
+    override fun onGetTemplate(): Template {
+        return MessageTemplate.Builder(carContext.getString(resId))
+            .setTitle(carContext.getString(R.string.app_name))
+            .setHeaderAction(Action.APP_ICON)
+            .setIcon(CarIcon.ALERT)
+            .addAction(
+                Action.Builder()
+                    .setTitle(carContext.getString(R.string.auto_retry))
+                    .setOnClickListener {
+                        screenManager.pop()
+                    }
+                    .build()
+            )
+            .build()
+    }
+}
